move add book mutation
