+++ mobile2/src/Environment.ts
@@ -3,7 +3,7 @@ import { Environment, Network, RecordSource, Store } from "relay-runtime";
 const store = new Store(new RecordSource());
 
 const network = Network.create((operation, variables) => {
-  return fetch("http://localhost:4000/graphql", {
+  return fetch("http://192.168.29.5:4000/graphql", {
     method: "POST",
     headers: {
       Accept: "application/json",
